git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@157244 4525493e-7705-40b1-a816-d608a930855b
Former-commit-id: 47368fdfab9de52293f5c5fe01ba7ad9e8cece03

+++ Online/Online/ROMon/ROMon/ROMon.h
@@ -40,13 +40,13 @@ namespace ROMon {
   PACK_DATA(class) MBMClient {
   public:
     /// Object name
-    char name[BM_USER_NAME_LEN];
-    int  partitionID;
-    int  processID;
-    int  events;
-    char reqs[4];
-    char type;
-    char state;
+    char  name[BM_USER_NAME_LEN];
+    int   partitionID;
+    int   processID;
+    long  events;
+    char  reqs[4];
+    char  type;
+    char  state;
     /// 3 Byte payload padding to ensure 32 bit alignment
     char pad[2];
   public:    // Public data accessors
